Exercise 12 + Finished notes!!
Nice lesson on how to create your own class (idk its more like variable to me) , and define attributes and methods yourself.

Finished the 'condensed' notes so I don't have to read through the long versions of 12 chapters later on during review.

Will do the 'documentary' later if I feel like it, but this is pretty much the end of the notebook.